Split cli-suite-builder file into suite and case
