+++ pkg/testing/cli-case.go
@@ -1,17 +1,3 @@
-// Copyright 2021 VMware
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package testing
 
 import (
@@ -71,41 +57,6 @@ const (
 	expectedDefaultFilename  = "expected.yaml"
 	yttValuesDefaultFilename = "ytt-values.yaml"
 )
-
-func buildTestSuite(testCase *TemplateTestCase, directory string) (TemplateTestSuite, error) {
-	var err error
-
-	testCase, err = populateTestCase(testCase, directory)
-	if err != nil {
-		return nil, fmt.Errorf("populate testCase: %w", err)
-	}
-
-	subdirectories, err := getSubdirectories(directory)
-	if err != nil {
-		return nil, fmt.Errorf("get subdirectories: %w", err)
-	}
-
-	// recurse
-	if len(subdirectories) > 0 {
-		testSuite := make(TemplateTestSuite)
-		for _, subdirectory := range subdirectories {
-			newCase := *testCase
-			var tempTestSuite TemplateTestSuite
-			tempTestSuite, err = buildTestSuite(&newCase, filepath.Join(directory, subdirectory))
-			if err != nil {
-				return nil, fmt.Errorf("failed building test case for subdirectory: %s: %w", subdirectory, err)
-			}
-			for name, aCase := range tempTestSuite {
-				testSuite[name] = aCase
-			}
-		}
-		return testSuite, nil
-	}
-
-	return TemplateTestSuite{
-		directory: testCase,
-	}, nil
-}
 
 func populateTestCase(testCase *TemplateTestCase, directory string) (*TemplateTestCase, error) {
 	info, err := populateInfo(directory)
@@ -227,21 +178,6 @@ func populateTestCaseMockSupplyChain(testCase *TemplateTestCase, info *testInfo)
 func populateTestCaseSupplyChain(testCase *TemplateTestCase, directory string, info *testInfo) (*TemplateTestCase, bool, error) {
 	panic("not implemented")
 	//return testCase, false, nil
-}
-
-func getSubdirectories(directory string) ([]string, error) {
-	var subdirectories []string
-	files, err := os.ReadDir(directory)
-	if err != nil {
-		return nil, fmt.Errorf("read dir: %w", err)
-	}
-
-	for _, file := range files {
-		if file.IsDir() {
-			subdirectories = append(subdirectories, file.Name())
-		}
-	}
-	return subdirectories, nil
 }
 
 func replaceIfFound(directory, filename string, priorityPath *string) (string, error) {
